Move deploy_data_nft to test_helpers.py
